Refactored javascript out of a single file Added webpack Added babel Added eslint
