+++ .eslintrc
@@ -1,0 +1,7 @@
+{
+  "parser": "babel-eslint",
+  "extends": "airbnb",
+  "rules": {
+    "comma-dangle": 0
+  }
+}
