+++ app/assets/javascripts/routes.js
@@ -1,0 +1,5 @@
+import subscribe from './subscribe';
+
+export default {
+  subscribe
+};
