+++ app/assets/stylesheets/repo.scss
@@ -1,0 +1,5 @@
+.settings {
+  .behaviors {
+    margin-top: 20px;
+  }
+}
