+++ app/routes.rb
@@ -1,4 +1,5 @@
 require_relative "routes/errors"
+require_relative "routes/dashboard"
 require_relative "routes/auth"
 require_relative "routes/actions"
 require_relative "routes/tasks"
@@ -7,3 +8,4 @@ require_relative "routes/webhooks"
 require_relative "routes/discoveries"
 require_relative "routes/users"
 require_relative "routes/account"
+require_relative "routes/content"
