+++ app/routes/repos.rb
@@ -16,7 +16,7 @@ module Pushbit
       authenticate!
       repo = repo_from_params
       authorize! :unsubscribe, repo
-      
+
       Cashier.unsubscribe(repo, current_user) if repo.private?
       Activator.deactivate(repo, current_user)
 
@@ -28,7 +28,7 @@ module Pushbit
       authenticate!
       repo = repo_from_params
       authorize! :trigger, repo
-      
+
       trigger = Trigger.create!(
         kind: "manual",
         repo: repo,
@@ -52,30 +52,30 @@ module Pushbit
       authorize! :read, repo
 
       @repo = repo
-      @tasks = @repo.tasks.paginate(page: params['page'])
-      @actions = Action.paginate(page: params['page']).where(repo_id: @repo.id).includes(:task, :user)
+      @behaviors = Behavior.all
       @title = @repo.github_full_name
 
       erb :'repos/show'
     end
 
-    get "/repos/:user/:repo/settings" do
+    get "/repos/:user/:repo/logs" do
       authenticate!
       repo = repo_from_params
       authorize! :update, repo
-      
-      @repo = repo
-      @behaviors = Behavior.all
-      @title = "Settings - #{@repo.github_full_name}"
 
-      erb :'repos/settings'
+      @repo = repo
+      @actions = Action.paginate(page: params['page']).where(repo_id: @repo.id).includes(:task, :user)
+      @tasks = @repo.tasks.paginate(page: params['page'])
+      @title = "Logs - #{@repo.github_full_name}"
+
+      erb :'repos/logs'
     end
 
     get "/repos/:user/:repo/:task_sequential_id" do
       authenticate!
       repo = repo_from_params
       authorize! :read, repo
-  
+
       @repo = repo
       @task = Task.find_by!(repo: @repo, sequential_id: params['task_sequential_id'])
       @actions = @task.actions.map { |a| ActionPresenter.new(a) }
