+++ app/views/_scripts.erb
@@ -12,4 +12,4 @@
   <% end %>
 </script>
 
-<%= javascript_tag 'app', :expand => true %>
+<%= javascript_tag 'bundle' %>
