+++ app/views/errors/500.erb
@@ -1,4 +1,4 @@
 <div class="vertically-centered">
   <h1>Error</h1>
-  <p class="subtitle">Sorry, an unexpected and embarrasing error occurred. Maybe try  <a href="javascript:window.location.reload()">that request</a> again?</p>
+  <p class="subtitle">Sorry, an unexpected and embarrassing error occurred. Maybe try  <a href="javascript:window.location.reload()">that request</a> again?</p>
 </div>
