+++ app/views/home.erb
@@ -10,7 +10,7 @@
         <li>Detect unoptimized images, compresses them and creates a new pull request.</li>
         <li>Find issues that need more information and ask the creator in a comment.</li>
         <li>Automatically assign a reviewer for new pull requests based on the code that changed.</li>
-        <li>Automatically merge pull requests when enough collaborators have given a thumbs up.</li>
+        <li>Merge pull requests when enough collaborators have given a thumbs up.</li>
       </ol>
     </div>
   </section>
