+++ app/views/layout.erb
@@ -8,24 +8,22 @@
     <meta name="csrf-token" content="<%= session[:csrf] %>">
 
     <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.5/css/bootstrap.min.css" integrity="sha512-dTfge/zgoMYpP7QbHy4gWMEGsbsdZeCXz7irItjcC3sPUFtf0kuFbDz/ixG7ArTxmDjLXDmezHubeNikyKGVyQ==" crossorigin="anonymous">
-    <script src="//code.jquery.com/jquery-1.11.3.min.js" crossorigin="anonymous"></script>
-
     <link rel="shortcut icon" type="image/png" href="<%= image_path 'favicon.png' %>">
     <link href='https://fonts.googleapis.com/css?family=Asap:400,700italic' rel='stylesheet' type='text/css'>
     <link rel="stylesheet" href="/octicons/octicons.css">
     <%= stylesheet_tag 'app' %>
 
     <script>
-  (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-  (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-  m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-  })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
+    (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
+    (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
+    m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
+    })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
 
-  ga('create', 'UA-1703053-21', 'auto');
-  ga('send', 'pageview');
+    ga('create', 'UA-1703053-21', 'auto');
+    ga('send', 'pageview');
     </script>
   </head>
-  <body>
+  <body id="<%= @id %>">
     <div class="wrapper">
       <%= partial :flash %>
       <header>
