+++ app/views/onboarding.erb
@@ -1,4 +1,4 @@
-<h1>Get Started</h1>
+<h1>Welcome</h1>
 <p>Pushbit works by checking out your pull requests, pushes, comments and other actions on Github to provide fixes and suggestions. Get started in just a few minutes by following these steps&hellip;</p>
 
 <div id="onboarding" class="row">
@@ -28,7 +28,7 @@
     </section>
     </li>
   </ol>
-  
+
   <form action="/users/me" class="col-md-12" id="skip-onboarding" method="POST">
     <input type="hidden" name="_method" value="PUT">
     <input type="hidden" name="onboarding_skipped" value="true" />
